refactor(v3): `cli-utils`: extract a `home-manager` module
This is the first commit that is intended to slowly remove Snowfall Lib
in favor of a simpler, homegrown approach.

My aim is to:
- Get rid of the rigid, uncomposable and over-complicated mess I've made
- Make the code simpler
- Reduce flake outputs (I really don't have a need for most of them,
and I suspect they add quite a bit of evaluation overhead)

The module in this commit is now a valid `home-manager` module -
though it's still imported using Snowfall machinery for the time
being. The plan is to use as many `home-manager` modules as possible, whilst
minimizing the number of nixos-modules.

I have some ideas on how to untagle some bits of this codebase in a
nice way.

Every module should remain *easily composable*. Copy-pasting at the
top-level for a system is *fine* and *encouraged*.